// web-frontend/src/app/libs/api/generated/orval/model/post.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Muslog API
 * OpenAPI spec version: 1.0.0
 */
import type { Track } from "./track";
import type { Tag } from "./tag";

export interface Post {
  id: number;
  title: string;
  description: string;
  userId: string;
  createdAt?: string;
  updatedAt?: string;
  tracks: Track[];
  tags?: Tag[];
  likesCount?: number;
}